finish

+++ database/migrations/2013_11_27_191653_create_gurus_table.php
@@ -15,7 +15,7 @@ class CreateGurusTable extends Migration
     {
         Schema::create('gurus', function (Blueprint $table) {
             $table->id();
-            $table->integer("nip");
+            $table->bigInteger("nip");
             $table->string("nama");
             $table->timestamps();
         });
